requirements.txt
langchain
langchain-core  
streamlit
python-dotenv  
langchain-huggingface
huggingface_hub
transformers
accelerate
bitsandbytes
langchain
langchain-community
langserve
fastapi
uvicorn
sse_starlette